+++ MagicalLifeGUIWindows/GUI/MainMenu/Buttons/QuitButton.cs
@@ -1,4 +1,5 @@
-﻿using MagicalLifeAPI.Universal;
+﻿using MagicalLifeAPI.Sound;
+using MagicalLifeAPI.Universal;
 using MagicalLifeGUIWindows.GUI.Reusable;
 using Microsoft.Xna.Framework;
 using MonoGame.Extended.Input.InputListeners;
@@ -13,12 +14,12 @@ namespace MagicalLifeGUIWindows.GUI.MainMenu.Buttons
 
         public override void Click(MouseEventArgs e)
         {
+            FMODUtil.RaiseEvent(EffectsTable.UIClick);
             UniversalEvents.GameExitHandler();
         }
 
         public override void DoubleClick(MouseEventArgs e)
         {
-            UniversalEvents.GameExitHandler();
         }
 
         private static Rectangle GetLocation()